Refactor review packet handoff context
